+++ packages/company/company.el
@@ -123,6 +123,14 @@
      :foreground "red"))
   "Face used for the selected common completion in the tooltip.")
 
+(defface company-tooltip-annotation
+  '((default :inherit company-tooltip)
+    (((background light))
+     :foreground "firebrick4")
+    (((background dark))
+     :foreground "red4"))
+  "Face used for the annotation in the tooltip.")
+
 (defface company-scrollbar-fg
   '((((background light))
      :background "darkred")
@@ -139,18 +147,27 @@
   "Face used for the tooltip scrollbar background.")
 
 (defface company-preview
-  '((t :background "blue4"
-       :foreground "wheat"))
+  '((((background light))
+     :inherit company-tooltip-selection)
+    (((background dark))
+     :background "blue4"
+     :foreground "wheat"))
   "Face used for the completion preview.")
 
 (defface company-preview-common
-  '((t :inherit company-preview
-       :foreground "red"))
+  '((((background light))
+     :inherit company-tooltip-selection)
+    (((background dark))
+     :inherit company-preview
+     :foreground "red"))
   "Face used for the common part of the completion preview.")
 
 (defface company-preview-search
-  '((t :inherit company-preview
-       :background "blue1"))
+  '((((background light))
+     :inherit company-tooltip-common-selection)
+    (((background dark))
+     :inherit company-preview
+     :background "blue1"))
   "Face used for the search string in the completion preview.")
 
 (defface company-echo nil
@@ -326,6 +343,13 @@ buffer with documentation for it.  Preferably use `company-doc-buffer',
 of buffer and buffer location, or of file and line number where the
 completion candidate was defined.
 
+`annotation': The second argument is a completion candidate.  Returns a
+string to be displayed inline with the candidate in the popup.  If
+duplicates are removed by company, candidates with equal string values will
+be kept if they have different annotations.  For that to work properly,
+backends should store the related information with candidates using text
+properties.
+
 `require-match': If this returns t, the user is not allowed to enter
 anything not offered as a candidate.  Use with care!  The default value nil
 gives the user that choice with `company-require-match'.  Return value
@@ -357,6 +381,15 @@ does not know about.  It should also be callable interactively and use
                            (symbol :tag "User defined"))))))
 
 (put 'company-backends 'safe-local-variable 'company-safe-backends-p)
+
+(defcustom company-transformers nil
+  "Functions to change the list of candidates received from backends,
+after sorting and removal of duplicates (if appropriate).
+Each function gets called with the return value of the previous one."
+  :type '(choice
+          (const :tag "None" nil)
+          (const :tag "Sort by occurrence" (company-sort-by-occurrence))
+          (repeat :tag "User defined" (function))))
 
 (defcustom company-completion-started-hook nil
   "Hook run when company starts completing.
@@ -441,7 +474,7 @@ immediately when a prefix of `company-minimum-prefix-length' is reached."
                  (const :tag "immediate (t)" t)
                  (number :tag "seconds")))
 
-(defcustom company-begin-commands '(self-insert-command)
+(defcustom company-begin-commands '(self-insert-command org-self-insert-command)
   "A list of commands after which idle completion is allowed.
 If this is t, it can show completions after any command.  See
 `company-idle-delay'.
@@ -625,6 +658,7 @@ means that `company-mode' is always turned on except in `message-mode' buffers."
 ;; Hack:
 ;; Emacs calculates the active keymaps before reading the event.  That means we
 ;; cannot change the keymap from a timer.  So we send a bogus command.
+;; XXX: Seems not to be needed anymore in Emacs 24.4
 (defun company-ignore ()
   (interactive)
   (setq this-command last-command))
@@ -768,8 +802,8 @@ Controlled by `company-auto-complete'.")
   (substring str (length company-prefix)))
 
 (defun company--insert-candidate (candidate)
+  (setq candidate (substring-no-properties candidate))
   ;; XXX: Return value we check here is subject to change.
-  (set-text-properties 0 (length candidate) nil candidate)
   (if (eq (company-call-backend 'ignore-case) 'keep-prefix)
       (insert (company-strip-prefix candidate))
     (delete-region (- (point) (length company-prefix)) (point))
@@ -864,10 +898,16 @@ can retrieve meta-data for them."
     (setq company-common
           (if (cdr company-candidates)
               (company--safe-candidate
-               (try-completion company-prefix company-candidates))
+               (let ((common (try-completion company-prefix company-candidates)))
+                 (if (eq common t)
+                     ;; Mulple equal strings, probably with different
+                     ;; annotations.
+                     company-prefix
+                   common)))
             (car company-candidates)))))
 
 (defun company--safe-candidate (str)
+  ;; XXX: This feature is deprecated.
   (or (company-call-backend 'crop str)
       str))
 
@@ -894,11 +934,8 @@ can retrieve meta-data for them."
           (unless (company-call-backend 'sorted)
             (setq candidates (sort candidates 'string<)))
           (when (company-call-backend 'duplicates)
-            ;; strip duplicates
-            (let ((c2 candidates))
-              (while c2
-                (setcdr c2 (progn (while (equal (pop c2) (car c2)))
-                                  c2)))))))
+            (company--strip-duplicates candidates))))
+    (setq candidates (company--transform-candidates candidates))
     (when candidates
       (if (or (cdr candidates)
               (not (eq t (compare-strings (car candidates) nil nil
@@ -906,6 +943,68 @@ can retrieve meta-data for them."
           candidates
         ;; Already completed and unique; don't start.
         t))))
+
+(defun company--strip-duplicates (candidates)
+  (let ((c2 candidates))
+    (while c2
+      (setcdr c2
+              (let ((str (car c2))
+                    (anno 'unk))
+                (pop c2)
+                (while (let ((str2 (car c2)))
+                         (if (not (equal str str2))
+                             nil
+                           (when (eq anno 'unk)
+                             (setq anno (company-call-backend
+                                         'annotation str)))
+                           (equal anno
+                                  (company-call-backend
+                                   'annotation str2))))
+                  (pop c2))
+                c2)))))
+
+(defun company--transform-candidates (candidates)
+  (let ((c candidates))
+    (dolist (tr company-transformers)
+      (setq c (funcall tr c)))
+    c))
+
+(defun company-sort-by-occurrence (candidates)
+  "Sort CANDIDATES according to their occurrences.
+Searches for each in the currently visible part of the current buffer and
+gives priority to the closest ones above point, then closest ones below
+point. The rest of the list is appended unchanged.
+Keywords and function definition names are ignored."
+  (let* (occurs
+         (noccurs
+          (delete-if
+           (lambda (candidate)
+             (when (or
+                    (save-excursion
+                      (progn (forward-line 0)
+                             (search-backward candidate (window-start) t)))
+                    (save-excursion
+                      (search-forward candidate (window-end) t)))
+               (let ((beg (match-beginning 0))
+                     (end (match-end 0)))
+                 (when (save-excursion
+                         (goto-char end)
+                         (and (not (memq (get-text-property (point) 'face)
+                                         '(font-lock-function-name-face
+                                           font-lock-keyword-face)))
+                              (let* ((prefix (company-call-backend 'prefix))
+                                     (prefix (or (car-safe prefix) prefix)))
+                                (and (stringp prefix)
+                                     (= (length prefix) (- end beg))))))
+                   (push (cons candidate (if (< beg (point))
+                                             (- (point) end)
+                                           (- beg (window-start))))
+                         occurs)
+                   t))))
+           candidates)))
+    (nconc
+     (mapcar #'car (sort occurs (lambda (e1 e2) (< (cdr e1) (cdr e2)))))
+     noccurs)))
 
 (defun company-idle-begin (buf win tick pos)
   (and company-mode
@@ -919,7 +1018,8 @@ can retrieve meta-data for them."
              (company-begin-commands t))
          (company-begin)
          (when company-candidates
-           (company-input-noop)
+           (when (version< emacs-version "24.3.50")
+             (company-input-noop))
            (company-post-command)))))
 
 (defun company-auto-begin ()
@@ -986,11 +1086,7 @@ can retrieve meta-data for them."
                                 company-point)
               company-prefix)))
 
-(defsubst company--string-incremental-p (old-prefix new-prefix)
-  (and (> (length new-prefix) (length old-prefix))
-       (equal old-prefix (substring new-prefix 0 (length old-prefix)))))
-
-(defun company--continue-failed (new-prefix)
+(defun company--continue-failed ()
   (when (company--incremental-p)
     (let ((input (buffer-substring-no-properties (point) company-point)))
       (cond
@@ -1001,8 +1097,7 @@ can retrieve meta-data for them."
           (let ((company--auto-completion t))
             (company-complete-selection))
           nil))
-       ((and (company--string-incremental-p company-prefix new-prefix)
-             (company-require-match-p))
+       ((company-require-match-p)
         ;; wrong incremental input, but required match
         (delete-char (- (length input)))
         (ding)
@@ -1042,7 +1137,7 @@ can retrieve meta-data for them."
           (setq company-prefix new-prefix)
           (company-update-candidates c)
           c)
-         (t (company--continue-failed new-prefix)))
+         (t (company--continue-failed)))
         (company-cancel))))
 
 (defun company--begin-new ()
@@ -1534,7 +1629,7 @@ To show the number next to the candidates in some back-ends, enable
 
 (defun company-fetch-metadata ()
   (let ((selected (nth company-selection company-candidates)))
-    (unless (equal selected (car company-last-metadata))
+    (unless (eq selected (car company-last-metadata))
       (setq company-last-metadata
             (cons selected (company-call-backend 'meta selected))))
     (cdr company-last-metadata)))
@@ -1715,20 +1810,16 @@ Example: \(company-begin-with '\(\"foo\" \"foobar\" \"foobarbaz\"\)\)"
       (pop copy))
     (apply 'concat pieces)))
 
-(defun company--highlight-common (line properties)
-  ;; XXX: Subject to change.
-  (let ((common (or (company-call-backend 'common-part line)
-                    (length company-common))))
-    (add-text-properties 0 common properties line)))
-
-(defun company-fill-propertize (line width selected)
-  (let* ((margin company-tooltip-margin)
-         (common (+ (or (company-call-backend 'common-part line)
-                        (length company-common)) margin)))
-    (setq line (concat (company-space-string company-tooltip-margin)
-                       (company-safe-substring
-                        line 0 (+ width company-tooltip-margin)))
-          width (+ width (* 2 margin)))
+(defun company-fill-propertize (value annotation width selected left right)
+  (let* ((margin (length left))
+         (common (+ (or (company-call-backend 'common-part value)
+                        (length company-common)) margin))
+         (ann-start (+ margin (length value)))
+         (line (concat left
+                       (company-safe-substring (concat value annotation)
+                                               0 width)
+                       right)))
+    (setq width (+ width margin (length right)))
 
     (add-text-properties 0 width '(face company-tooltip
                                    mouse-face company-tooltip-mouse)
@@ -1737,16 +1828,20 @@ Example: \(company-begin-with '\(\"foo\" \"foobar\" \"foobarbaz\"\)\)"
                          '(face company-tooltip-common
                            mouse-face company-tooltip-mouse)
                          line)
+    (add-text-properties ann-start (+ ann-start (length annotation))
+                         '(face company-tooltip-annotation
+                           mouse-face company-tooltip-mouse)
+                         line)
     (when selected
       (if (and company-search-string
-               (string-match (regexp-quote company-search-string) line
+               (string-match (regexp-quote company-search-string) value
                              (length company-prefix)))
-          (progn
-            (add-text-properties (match-beginning 0) (match-end 0)
-                                 '(face company-tooltip-selection)
+          (let ((beg (+ margin (match-beginning 0)))
+                (end (+ margin (match-end 0))))
+            (add-text-properties beg end '(face company-tooltip-selection)
                                  line)
-            (when (< (match-beginning 0) common)
-              (add-text-properties (match-beginning 0) common
+            (when (< beg common)
+              (add-text-properties beg common
                                    '(face company-tooltip-common-selection)
                                    line)))
         (add-text-properties 0 width '(face company-tooltip-selection
@@ -1755,8 +1850,8 @@ Example: \(company-begin-with '\(\"foo\" \"foobar\" \"foobarbaz\"\)\)"
         (add-text-properties margin common
                              '(face company-tooltip-common-selection
                                mouse-face company-tooltip-selection)
-                             line))))
-  line)
+                             line)))
+    line))
 
 ;;; replace
 
@@ -1829,17 +1924,16 @@ Example: \(company-begin-with '\(\"foo\" \"foobar\" \"foobarbaz\"\)\)"
     line))
 
 (defun company--create-lines (selection limit)
-
   (let ((len company-candidates-length)
         (numbered 99999)
         (window-width (company--window-width))
         lines
         width
         lines-copy
+        items
         previous
         remainder
-        scrollbar-bounds
-        new)
+        scrollbar-bounds)
 
     ;; Maybe clear old offset.
     (when (<= len (+ company-tooltip-offset limit))
@@ -1871,40 +1965,49 @@ Example: \(company-begin-with '\(\"foo\" \"foobar\" \"foobarbaz\"\)\)"
     (when scrollbar-bounds (decf window-width))
 
     (dotimes (_ len)
-      (setq width (max (length (pop lines-copy)) width)))
+      (let* ((value (pop lines-copy))
+             (annotation (company-call-backend 'annotation value)))
+        (push (cons value annotation) items)
+        (setq width (max (+ (length value) (length annotation)) width))))
+
     (setq width (min window-width
                      (if (and company-show-numbers
                               (< company-tooltip-offset 10))
                          (+ 2 width)
                        width)))
-    (setq lines-copy lines)
 
     ;; number can make tooltip too long
     (when company-show-numbers
       (setq numbered company-tooltip-offset))
 
-    (when previous
-      (push (company--scrollpos-line previous width) new))
+    (let ((items (nreverse items)) new)
+      (when previous
+        (push (company--scrollpos-line previous width) new))
 
-    (dotimes (i len)
-      (let ((line (company-fill-propertize
-                   (if (>= numbered 10)
-                       (company-reformat (pop lines))
-                     (incf numbered)
-                     (format "%s %d"
-                             (company-safe-substring
-                              (company-reformat (pop lines)) 0 (- width 2))
-                             (mod numbered 10)))
-                   width (equal i selection))))
-        (push (if scrollbar-bounds
-                  (company--scrollbarize line i scrollbar-bounds)
-                line)
-              new)))
+      (dotimes (i len)
+        (let* ((item (pop items))
+               (str (company-reformat (car item)))
+               (annotation (cdr item))
+               (right (company-space-string company-tooltip-margin))
+               (width width))
+          (when (< numbered 10)
+            (decf width 2)
+            (incf numbered)
+            (setq right (concat (format " %d" (mod numbered 10)) right)))
+          (push (concat
+                 (company-fill-propertize str annotation
+                                          width (equal i selection)
+                                          (company-space-string
+                                           company-tooltip-margin)
+                                          right)
+                 (when scrollbar-bounds
+                   (company--scrollbar i scrollbar-bounds)))
+                new)))
 
-    (when remainder
-      (push (company--scrollpos-line remainder width) new))
+      (when remainder
+        (push (company--scrollpos-line remainder width) new))
 
-    (setq lines (nreverse new))))
+      (nreverse new))))
 
 (defun company--scrollbar-bounds (offset limit length)
   (when (> length limit)
@@ -1913,12 +2016,11 @@ Example: \(company-begin-with '\(\"foo\" \"foobar\" \"foobarbaz\"\)\)"
            (upper (+ lower size -1)))
       (cons lower upper))))
 
-(defun company--scrollbarize (line i bounds)
-  (concat line
-          (propertize " " 'face
-                      (if (and (>= i (car bounds)) (<= i (cdr bounds)))
-                          'company-scrollbar-fg
-                        'company-scrollbar-bg))))
+(defun company--scrollbar (i bounds)
+  (propertize " " 'face
+              (if (and (>= i (car bounds)) (<= i (cdr bounds)))
+                  'company-scrollbar-fg
+                'company-scrollbar-bg)))
 
 (defun company--scrollpos-line (text width)
   (propertize (concat (company-space-string company-tooltip-margin)
@@ -2065,9 +2167,9 @@ Returns a negative number if the tooltip should be displayed above point."
 (defun company-preview-show-at-point (pos)
   (company-preview-hide)
 
-  (setq company-preview-overlay (make-overlay pos pos))
+  (setq company-preview-overlay (make-overlay pos (1+ pos)))
 
-  (let ((completion(nth company-selection company-candidates)))
+  (let ((completion (nth company-selection company-candidates)))
     (setq completion (propertize completion 'face 'company-preview))
     (add-text-properties 0 (length company-common)
                          '(face company-preview-common) completion)
@@ -2086,7 +2188,9 @@ Returns a negative number if the tooltip should be displayed above point."
          (not (equal completion ""))
          (add-text-properties 0 1 '(cursor t) completion))
 
-    (overlay-put company-preview-overlay 'after-string completion)
+    (overlay-put company-preview-overlay 'display
+                 (concat completion (unless (eq pos (point-max))
+                                      (buffer-substring pos (1+ pos)))))
     (overlay-put company-preview-overlay 'window (selected-window))))
 
 (defun company-preview-hide ()